Move vertify class part
